requirements.txt
requests==2.23.0
docopt-ng==0.7.2
pygame==1.9.6
